pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-coder"
version = "0.1.0"
description = "Autonomous coding agent orchestrator"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.78.0",
    "pyyaml>=6.0",
]

[project.scripts]
auto-coder = "auto_coder.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["auto_coder*"]
